debug

+++ include/DSM.h
@@ -179,7 +179,7 @@ public:
                        value, node_num);
   }
 
-  uint64_t max_total(double value, int node_num) {
+  uint64_t max_total(uint64_t value, int node_num) {
     static uint64_t count = 0;
     return keeper->max(std::string("max-total") + std::to_string(count++),
                        value, node_num);
