+++ src/DSMKeeper.cpp
@@ -248,7 +248,7 @@ uint64_t DSMKeeper::min(const std::string &min_key, uint64_t value,
   return min;
 }
 
-uint64_t DSMKeeper::max(const std::string &max_key, double value,
+uint64_t DSMKeeper::max(const std::string &max_key, uint64_t value,
                         int node_num) {
   std::string key_prefix = std::string("max-") + max_key;
 
@@ -256,12 +256,12 @@ uint64_t DSMKeeper::max(const std::string &max_key, double value,
   memSet(key.c_str(), key.size(), (char *)&value, sizeof(value));
 
   uint64_t ret = 0;
-  double max = 0;
+  uint64_t max = 0;
   for (int i = 0; i < node_num; ++i) {
     key = key_prefix + std::to_string(i);
     auto mem_ret = memGet(key.c_str(), key.size(), nullptr, false);
     if (mem_ret != nullptr) {
-      auto cur_val = *(double *)(mem_ret);
+      auto cur_val = *(uint64_t *)(mem_ret);
       if (cur_val > max) {
         max = cur_val;
       }
